+++ config/options.js
@@ -1,0 +1,4 @@
+global.boundaries = {
+    top: 900,
+    left: 1500
+}